requirements.txt
streamlit
torch
torchvision
timm
Pillow
numpy
opencv-python
ultralytics
easyocr
transformers
whisper
ffmpeg-python